+++ ModelSwapperSkins/BoneMapping/BoneInitializer.cs
@@ -47,6 +47,7 @@ namespace ModelSwapperSkins.BoneMapping
             AddCustomBoneInitializerRules(BoneInitializerRules_Brother.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Captain.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Chef.Instance);
+            AddCustomBoneInitializerRules(BoneInitializerRules_Child.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Clay.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_ClayBruiser.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_ClayGrenadier.Instance);